+++ app/src/main/java/app/pachli/StatusListActivity.kt
@@ -28,9 +28,9 @@ import app.pachli.appstore.EventHub
 import app.pachli.appstore.FilterChangedEvent
 import app.pachli.components.compose.ComposeActivity
 import app.pachli.components.timeline.TimelineFragment
-import app.pachli.components.timeline.TimelineKind
 import app.pachli.core.network.model.Filter
 import app.pachli.core.network.model.FilterV1
+import app.pachli.core.network.model.TimelineKind
 import app.pachli.databinding.ActivityStatuslistBinding
 import app.pachli.interfaces.ActionButtonActivity
 import app.pachli.interfaces.AppBarLayoutHost
