+++ app/src/main/java/app/pachli/components/timeline/NetworkTimelineRepository.kt
@@ -28,6 +28,7 @@ import app.pachli.components.timeline.viewmodel.NetworkTimelineRemoteMediator
 import app.pachli.components.timeline.viewmodel.PageCache
 import app.pachli.core.accounts.AccountManager
 import app.pachli.core.network.model.Status
+import app.pachli.core.network.model.TimelineKind
 import app.pachli.core.network.retrofit.MastodonApi
 import app.pachli.util.getDomain
 import kotlinx.coroutines.CoroutineScope
