+++ app/src/main/java/app/pachli/components/timeline/viewmodel/NetworkTimelineRemoteMediator.kt
@@ -23,9 +23,9 @@ import androidx.paging.LoadType
 import androidx.paging.PagingState
 import androidx.paging.RemoteMediator
 import app.pachli.BuildConfig
-import app.pachli.components.timeline.TimelineKind
 import app.pachli.core.accounts.AccountManager
 import app.pachli.core.network.model.Status
+import app.pachli.core.network.model.TimelineKind
 import app.pachli.core.network.retrofit.MastodonApi
 import kotlinx.coroutines.CoroutineScope
 import retrofit2.HttpException
