+++ app/src/test/java/app/pachli/components/timeline/NetworkTimelineRemoteMediatorTest.kt
@@ -31,6 +31,7 @@ import app.pachli.components.timeline.viewmodel.PageCache
 import app.pachli.core.accounts.AccountManager
 import app.pachli.core.database.model.AccountEntity
 import app.pachli.core.network.model.Status
+import app.pachli.core.network.model.TimelineKind
 import com.google.common.truth.Truth.assertThat
 import kotlinx.coroutines.test.runTest
 import okhttp3.Headers
